fix: 폴더 구조 및 파일명 컨벤션 수정
