+++ app/pages/main/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./github-server";
